Move navbar file
