+++ src/components/layout/Layout.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from '../navbar/Navbar';
+import Navbar from './Navbar';
 
 export default function Layout({ children }) {
   return(
